+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kf1769gxxj0kehtdw71454br"
+        "conv_01kf2rj60tt3xx0jq2ktmwm5xx"
     )
     r"""Unique conversation identifier with `conv_` prefix."""