+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF1769YNG8BE2C84HH2E9MQT"
+        "01KF2RJ6E96B63E6FNN6ZEAS13"
     )
     r"""The unique identifier of the data source"""