+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kf1769haa390dm6qvw0tjk3h"
+        "conv_01kf2rj611tq907jg9yxkv8sw4"
     )
     r"""Unique conversation identifier with `conv_` prefix."""