+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF1769YR6RNWGT41J50QAZY9"
+        "01KF2RJ6ECKYMCBES2N69WY99V"
     )
     r"""The unique identifier of the data source"""